pyproject.toml
[tool.poetry]
name = "w2w"
version = "0.1"
description = ""
authors = ["posholokia <ilya.posholokk@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
djangorestframework = "^3.15.1"
gunicorn = "^22.0.0"
uvicorn = {extras = ["standard"], version = "^0.30.1"}
drf-spectacular = "^0.27.2"
pillow = "^10.3.0"
redis = "^5.0.5"
pytelegrambotapi = "^4.19.1"
python-dotenv = "^1.0.1"
djangorestframework-simplejwt = "^5.3.1"
channels = {extras = ["daphne"], version = "^4.1.0"}
djangochannelsrestframework = "^1.2.0"
channels-redis = "^4.2.0"
pyjwt = "^2.8.0"


[tool.poetry.group.win.dependencies]
psycopg2 = "^2.9.9"

[tool.poetry.group.dev.dependencies]
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
